✅ Test uint128s

+++ test/Uint256/BaseQuantoPerUSDUint256.t.sol
@@ -15,7 +15,7 @@ contract BaseQuantoPerUSDUint256Test is Test {
     }
 
     function testBaseQuantoPerUSDUint256AddFuzz(uint256 x, uint256 y) public {
-        uint z;
+        uint256 z;
         assembly {
             z := add(x, y)
         }
@@ -56,7 +56,7 @@ contract BaseQuantoPerUSDUint256Test is Test {
     }
 
     function testBaseQuantoPerUSDUint256AndFuzz(uint256 x, uint256 y) public {
-        uint z = x & y;
+        uint256 z = x & y;
         BaseQuantoPerUSDUint256 result = BaseQuantoPerUSDUint256.wrap(x).and(y);
         assertEq(result.unwrap(), z);
     }
@@ -69,7 +69,7 @@ contract BaseQuantoPerUSDUint256Test is Test {
     }
 
     function testBaseQuantoPerUSDUint256And2Fuzz(uint256 x, uint256 y) public {
-        uint z = x & y;
+        uint256 z = x & y;
         BaseQuantoPerUSDUint256 result = BaseQuantoPerUSDUint256.wrap(x) &
             BaseQuantoPerUSDUint256.wrap(y);
         assertEq(result.unwrap(), z);
@@ -167,7 +167,7 @@ contract BaseQuantoPerUSDUint256Test is Test {
             vm.expectRevert();
             BaseQuantoPerUSDUint256.wrap(x) % BaseQuantoPerUSDUint256.wrap(y);
         } else {
-            uint z = x % y;
+            uint256 z = x % y;
             BaseQuantoPerUSDUint256 result = BaseQuantoPerUSDUint256.wrap(x) %
                 BaseQuantoPerUSDUint256.wrap(y);
             assertEq(result.unwrap(), z);
@@ -191,7 +191,7 @@ contract BaseQuantoPerUSDUint256Test is Test {
     }
 
     function testBaseQuantoPerUSDUint256Not() public {
-        uint hundred = 100;
+        uint256 hundred = 100;
         BaseQuantoPerUSDUint256 x = BaseQuantoPerUSDUint256.wrap(hundred);
         BaseQuantoPerUSDUint256 result = ~x;
         uint256 expected = ~hundred;
@@ -236,7 +236,7 @@ contract BaseQuantoPerUSDUint256Test is Test {
     }
 
     function testBaseQuantoPerUSDUint256IncrementFuzz(uint256 x) public {
-        uint z;
+        uint256 z;
         assembly {
             z := add(x, 1)
         }
@@ -259,7 +259,7 @@ contract BaseQuantoPerUSDUint256Test is Test {
     }
 
     function testBaseQuantoPerUSDUint256MulFuzz(uint256 x, uint256 y) public {
-        uint z;
+        uint256 z;
         assembly {
             z := mul(x, y)
         }
@@ -285,7 +285,7 @@ contract BaseQuantoPerUSDUint256Test is Test {
         uint256 x,
         uint256 y
     ) public {
-        uint z;
+        uint256 z;
         assembly {
             z := div(
                 mul(x, y),
@@ -318,7 +318,7 @@ contract BaseQuantoPerUSDUint256Test is Test {
         uint256 x,
         uint256 y
     ) public {
-        uint z;
+        uint256 z;
         assembly {
             z := div(
                 mul(x, y),
@@ -353,7 +353,7 @@ contract BaseQuantoPerUSDUint256Test is Test {
         uint256 x,
         uint256 y
     ) public {
-        uint z;
+        uint256 z;
         assembly {
             z := div(
                 mul(x, y),
@@ -385,7 +385,7 @@ contract BaseQuantoPerUSDUint256Test is Test {
     }
 
     function testBaseQuantoPerUSDUint256DivFuzz(uint256 x, uint256 y) public {
-        uint z;
+        uint256 z;
         assembly {
             z := div(x, y)
         }
